requirements.txt
flask>=3.0
mutagen>=1.45
pillow>=10.0
waitress>=3.0
pywebview>=6.0
